[css-grid] Use max size to compute auto repeat tracks
When available size is indefinite we should use max size to compute
the number of auto repeat tracks.

The spec text is very clear
(https://drafts.csswg.org/css-grid/#auto-repeat):
> When auto-fill is given as the repetition number,
> if the grid container has a definite size or **max size**
> in the relevant axis...

So far we were not doing that for widths,
in this patch we modify LayoutGrid::ComputeAutoRepeatTracksCount()
to do the same than for heights.

BUG=957932
TEST=external/wpt/css/css-grid/grid-definition/grid-auto-repeat-max-size-001.html

Change-Id: I9cce24d38f1a00bea8fa04e42645c777c57e486e