[Background Sync] Add 'periodic-background-sync' permission.
This CL adds a permission and a feature flag for Periodic Background Sync.
The permission will be denied when the feature is disabled, which is true by
default while we're implementing it. We currently only grant the permission
if accessed from a secure origin.

Unit tests and a web test have also been added.

Future CLs will add more logic to PeriodicBackgroundSyncPermissionContext.

Bug: 925297
Change-Id: I41e8d692a8bea9c356dad8681a8c79e23d7eda4a