Monkey-patched ICE connection "failed" state to "disconnected".
This ensures that the terminal "failed" state isn't entered
prematurely on the PC level.

Added tests for ICEConnectionState interactions with candidates.

This should test interactions with/without a=trickle-ice, with/without
bogus candidates, late added candidates, in-offer candidates and so on.
The current test is for what happens when trickle is negotiated,
but the remote candidate given doesn't connect.

Bug: chromium:957487
Change-Id: I4554d7b4a24288b405bed2b1042ab601717220ca